backend/requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0

# Environment & Configuration
python-dotenv==1.0.0

# Validation & Data Models
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development Tools
black==23.12.0
flake8==7.0.0
autopep8==2.0.4

# Future Features (Commented out - uncomment when implementing)
# PyPDF2==3.0.1              # For PDF text extraction
# pdfplumber==0.10.3         # Alternative PDF extractor
# newspaper3k==0.2.8         # For URL article extraction
# beautifulsoup4==4.12.2     # For web scraping
# requests==2.31.0           # For HTTP requests
# readability-lxml==0.8.1    # For article extraction

# Future Database (Commented out - uncomment when implementing)
# SQLAlchemy==2.0.23
# Flask-SQLAlchemy==3.1.1
# Flask-Migrate==4.0.5

# Future Rate Limiting (Commented out - uncomment when implementing)
# Flask-Limiter==3.5.0

# Future Caching (Commented out - uncomment when implementing)
# Flask-Caching==2.1.0
# redis==5.0.1
